back to floats

+++ src/kernals/vector_add.cpp
@@ -16,11 +16,11 @@ extern "C" void AddVecs(float *d_dest, const float *d_a, const float *d_b, size_
     K_VectorAdd<<<num_blocks, ThreadsPerBlock>>>(d_dest, d_a, d_b, len);
 }
 
-__global__ void K_SqMatMul(int32_t *dest, const int32_t *a, const int32_t *b, size_t n) {
+__global__ void K_SqMatMul(float *dest, const float *a, const float *b, size_t n) {
     size_t col = blockIdx.x * blockDim.x + threadIdx.x;
     size_t row = blockIdx.y * blockDim.y + threadIdx.y;
 
-    int32_t dotProduct = 0;
+    float dotProduct = 0;
     if((col < n) && (row < n)) {
         for(size_t i = 0; i < n; i++) {
             // mat[r][c] = mat[r*n + c]
@@ -31,7 +31,7 @@ __global__ void K_SqMatMul(int32_t *dest, const int32_t *a, const int32_t *b, si
     }
 }
 
-extern "C" void SqMatMul(int32_t *d_dest, const int32_t *d_a, const int32_t *d_b, size_t n) {
+extern "C" void SqMatMul(float *d_dest, const float *d_a, const float *d_b, size_t n) {
     size_t blockDimX = sqrt(ThreadsPerBlock);
     size_t gridDimX = (n + blockDimX - 1) / blockDimX;
     dim3 dimBlock(blockDimX, blockDimX);
